Refactor document pipeline into reusable service
